新增登录、登出、注册

+++ userservice/src/main/java/cn/edu/xmu/user/model/vo/NewUserVo.java
@@ -5,6 +5,7 @@ import org.hibernate.validator.constraints.Length;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.Pattern;
+import java.time.LocalDateTime;
 
 @Data
 public class NewUserVo {
@@ -22,7 +23,7 @@ public class NewUserVo {
     @Pattern(regexp="[+]?[0-9*#]+",message="手机号格式不正确")
     private String mobile;
 
-    private int gender;
+    private Byte gender;
 
     private String birthday;
 }
